pip_requirements.txt
Django==1.9.2
Pillow==3.1.1
PyAudio==0.2.9
SpeechRecognition==3.2.0
argparse==1.2.1
image==1.4.4
numpy==1.10.4
wsgiref==0.1.2
zbar==0.10
